개인정보, 폼제출 막는 기능, 예약자동일, 삭제 confirm기능

+++ BeFourAfter/src/main/webapp/WEB-INF/views/reserve/list.jsp
@@ -89,7 +89,7 @@
 						</c:otherwise>
 					</c:choose>
 					<td><a href="${pageContext.request.contextPath }/reserve/updateform.do?num=${ tmp.num}">일정 수정</a></td>
-					<td><a href="${pageContext.request.contextPath }/reserve/delete.do?num=${ tmp.num}">삭제</a></td>
+					<td><a href="${pageContext.request.contextPath }/reserve/delete.do?num=${ tmp.num}" onclick="return confirm('예약을 취소하시겠습니까?')">삭제</a></td>
 				</tr>
 				</c:forEach>
 			</tbody>
